backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.100.0
anyio==4.13.0
asyncpg==0.31.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==48.0.0
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.1
filelock==3.29.0
fsspec==2026.4.0
h11==0.16.0
hf-xet==1.5.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.14.0
idna==3.14
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
markdown-it-py==4.2.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
openai==2.36.0
packaging==26.2
pdfminer.six==20251230
pdfplumber==0.11.9
pgvector==0.4.2
pillow==12.2.0
pycparser==3.0
pydantic==2.13.4
pydantic_core==2.46.4
Pygments==2.20.0
PyMuPDF==1.27.2.3
pypdfium2==5.8.0
python-dotenv==1.2.2
python-multipart==0.0.28
PyYAML==6.0.3
regex==2026.5.9
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
sniffio==1.3.1
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.8.0
typer==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
